scraper/requirements.txt
# Job Scraper Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
firebase-admin==6.2.0
schedule==1.2.1
zenrows==1.3.0
httpx==0.25.2
